requirements.txt
mediapipe==0.10.9
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.72
pandas==2.2.0
python-dotenv==1.0.1
scikit_learn==1.3.2
torch==2.1.2
djitellopy==2.5.0
pygame==2.5.2
tenacity==8.2.3
huggingface-hub==0.21.4
